pyproject.toml
[project]
name = "week-08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.157.0",
    "crewai-tools>=0.60.0",
    "openai>=1.99.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "tiktoken>=0.11.0",
    "pysqlite3-binary; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "rich>=13.9.4",
    "langchain-experimental>=0.3.4",
    "chromadb>=0.5.23",
    "bs4>=0.0.2",
]
